+++ langchain4j-alloydb-pg/src/main/java/dev/langchain4j/index/IVFFlatIndex.java
@@ -1,36 +1,33 @@
-package dev.langchain4j.store.alloydb.index;
+package dev.langchain4j.index;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class IVFIndex implements VectorIndex {
+public class IVFFlatIndex implements VectorIndex {
 
     private final Integer listCount;
     private final Integer probes;
-    private final String quantizer;
     private final DistanceStrategy distanceStrategy;
 
-    public IVFIndex(Integer listCount, Integer probes, String quantizer, DistanceStrategy distanceStrategy) {
-        this.listCount = (listCount != null) ? listCount : 100;
+    public IVFFlatIndex(Integer listCount, Integer probes, DistanceStrategy distanceStrategy) {
+        this.listCount = (listCount != null) ? listCount : 1;
         this.probes = (probes != null) ? probes : 1;
-        this.quantizer = (quantizer != null) ? quantizer : "sq8";
         this.distanceStrategy = distanceStrategy != null ? distanceStrategy : DistanceStrategy.DistanceStrategyFactory.getCosineDistanceStrategy();
     }
 
     @Override
     public String getIndexOptions() {
-        return String.format("(lists = %s, quantizer = %s)", listCount, quantizer);
+        return String.format("(lists = %s)", listCount);
     }
 
     @Override
     public List<String> getParameterSettings() {
         List<String> parameters = new ArrayList();
-        parameters.add(String.format("ivf.probes = %d", probes));
+        parameters.add(String.format("ivfflat.probes = %d", probes));
         return parameters;
     }
 
     public DistanceStrategy getDistanceStrategy() {
         return distanceStrategy;
     }
-
 }
